Adjust namespaces
