+++ DbSeeder/Data/Sql/BulkInsertSqlQuery.cs
@@ -1,0 +1,8 @@
+namespace DbSeeder.Data.Sql;
+
+public record BulkInsertSqlQuery(
+    string Table,
+    IReadOnlyList<IReadOnlyDictionary<string, string>> Values) // [col: value]
+{
+
+}
